improve query scoring
Remove the strong boost on name scoring. Introduce an additional
match against the ngram filter on names instead which ensures that
even partial matches lead to a certain boost.

Importance now boosts the name and housenumber matching instead of
the whole query. This avoids that places with high importance pop
up even when the query only matches against its address.